Update net worth type, expand README, add diagram
Changed userEstimatedNetWorth from double to decimal for better financial precision in Program.cs. Rewrote and expanded README.md with detailed documentation, diagrams, and test cases. Added diagram.png for project visuals.

+++ BasicVariables/Program.cs
@@ -28,7 +28,7 @@ namespace BasicVariables
             Console.Write("Are you a permanent resident? (true/false): ");
             bool userIsPermanentResident = bool.Parse(Console.ReadLine());
             Console.Write("Please enter your estimated net worth:");
-            double userEstimatedNetWorth = double.Parse(Console.ReadLine());
+            decimal userEstimatedNetWorth = decimal.Parse(Console.ReadLine());
             Console.Write("Please enter your LinkedIn profile URL: ");
             String userLinkedin = Console.ReadLine();
             Console.WriteLine("");
